Set restart policy for systemd unit
Currently, no restart policy is set. This means,
when node_exporter is killed for some reason, it
will not restart. Setting Restart=on-failure
mitigates this problem. This behaviour is
configurable via a variable, default is set to
on-failure.

The same policy is set in the official debian
package: https://salsa.debian.org/go-team/packages/prometheus-node-exporter/-/blob/debian/sid/debian/service#L6